New pool status statistics, highlighting in various place
Added new pool status display to the dashboard (Issue #21). Added rejected share
percentage to the worker status table. Added a new config setting for both
tables to highlight rows when their rejection percentage is at the specified
percentage. Added new highlighting to the worker status table when the worker
has not submitted any shares in the amount of seconds specified in
average_interval setting in config.